don't need ip_address for example server
