+++ responses/src/available_encodings.rs
@@ -1,10 +1,4 @@
-#[derive(Clone, Debug)]
-pub struct AvailableEncodings {
-    gzip: bool,
-    deflate: bool,
-    br: bool,
-    zstd: bool,
-}
+use crate::type_flyweight::AvailableEncodings;
 
 impl AvailableEncodings {
     pub fn new(potential_encodings: &Vec<String>) -> AvailableEncodings {
@@ -15,17 +9,21 @@ impl AvailableEncodings {
             zstd: false,
         };
 
+        av_enc.update(potential_encodings);
+
+        av_enc
+    }
+
+    pub fn update(&mut self, potential_encodings: &Vec<String>) {
         for encoding in potential_encodings {
             match encoding.as_str() {
-                "gzip" => av_enc.gzip = true,
-                "deflate" => av_enc.deflate = true,
-                "br" => av_enc.br = true,
-                "zstd" => av_enc.zstd = true,
+                "gzip" => self.gzip = true,
+                "deflate" => self.deflate = true,
+                "br" => self.br = true,
+                "zstd" => self.zstd = true,
                 _ => {}
             }
         }
-
-        av_enc
     }
 
     pub fn encoding_is_available(&self, encoding: &str) -> bool {
