+++ responses/src/lib.rs
@@ -1,8 +1,11 @@
+mod available_encodings;
 mod content_type;
 mod request_details;
 mod responses;
+mod service_requirements;
 mod type_flyweight;
 
 pub use crate::request_details::get_request_details;
 pub use crate::responses::build_response;
-pub use crate::type_flyweight::BoxedResponse;
+pub use crate::service_requirements::get_service_requirements;
+pub use crate::type_flyweight::{AvailableEncodings, BoxedResponse, ServiceRequirements};
