+++ server/src/service.rs
@@ -4,33 +4,28 @@ use hyper::Request;
 use std::future::Future;
 use std::pin::Pin;
 
-use responses;
+/*
+    BoxedResponse is a type.
+    It should work with hyper responses across
+    different libraries and dependencies.
+*/
 
-use config::AvailableEncodings;
-use config::ServiceRequirements;
+use responses::{AvailableEncodings, BoxedResponse, ServiceRequirements};
 
 pub struct Svc {
     service_requirements: ServiceRequirements,
-    available_encodings: AvailableEncodings,
-    ip_address: String,
 }
 
 impl Svc {
-    pub fn new(
-        service_requirements: &ServiceRequirements,
-        available_encodings: &AvailableEncodings,
-        ip_address: &str,
-    ) -> Svc {
+    pub fn new(service_requirements: &ServiceRequirements) -> Svc {
         Svc {
             service_requirements: service_requirements.clone(),
-            available_encodings: available_encodings.clone(),
-            ip_address: ip_address.to_string(),
         }
     }
 }
 
 impl Service<Request<IncomingBody>> for Svc {
-    type Response = responses::BoxedResponse;
+    type Response = BoxedResponse;
     type Error = hyper::http::Error;
     type Future = Pin<Box<dyn Future<Output = Result<Self::Response, Self::Error>> + Send>>;
